+++ tests/integration_tests/reference-tests/multiple-packages/se/repomate/fibo/FiboTest.java
