+++ tests/integration_tests/test_end_to_end.py
@@ -1,4 +1,4 @@
-"""
+"""End-to-end tests for the plugin.
 
 .. important::
 
@@ -9,24 +9,18 @@
         variable REPOMATE_JUNIT4_HAMCREST.
         3. A path to ``junit-4.12.jar`` must be in the environment variable
         REPOMATE_JUNIT4_JUNIT.
-
 """
-
 import pathlib
 import shutil
 import tempfile
 import os
-import re
 from configparser import ConfigParser
 from collections import namedtuple
-from argparse import ArgumentParser
 
 import pytest
 
-import repomate_plug as plug
 from repomate_plug import Status
 from repomate_junit4 import junit4
-from repomate_junit4 import _junit4_runner
 
 import envvars
 
@@ -377,169 +371,3 @@ class TestSecurityPolicy:
 
         assert result.status == Status.SUCCESS
         assert "Test class FiboTest passed!" in result.msg
-
-
-class TestParseArgs:
-    def test_all_args(self, junit4_hooks, full_args):
-        """Test that args-related attributes are correctly set when all of them
-        are in the args namespace.
-        """
-        junit4_hooks.parse_args(full_args)
-
-        assert junit4_hooks._master_repo_names == MASTER_REPO_NAMES
-        assert junit4_hooks._reference_tests_dir == RTD
-        assert junit4_hooks._ignore_tests == IGNORE_TESTS
-        assert junit4_hooks._hamcrest_path == HAMCREST_PATH
-        assert junit4_hooks._junit_path == JUNIT_PATH
-        assert junit4_hooks._disable_security == False
-
-    def test_defaults_are_overwritten(self, junit4_hooks, full_args):
-        """Test that ignore_tests, hamcrest_path and junit_path are all
-        overwritten by the parse_args method, even if they are set to something
-        previously.
-        """
-        junit4_hooks._ignore_tests = "this isn't even a list"
-        junit4_hooks._hamcrest_path = "wrong/path"
-        junit4_hooks._junit_path = "also/wrong/path"
-        junit4_hooks._reference_tests_dir = "some/cray/dir"
-
-        junit4_hooks.parse_args(full_args)
-
-        assert junit4_hooks._ignore_tests == IGNORE_TESTS
-        assert junit4_hooks._hamcrest_path == HAMCREST_PATH
-        assert junit4_hooks._junit_path == JUNIT_PATH
-        assert junit4_hooks._reference_tests_dir == RTD
-
-    def test_defaults_are_kept_if_not_specified_in_args(self, junit4_hooks, full_args):
-        """Test that defaults are not overwritten if they are falsy in the
-        args.
-        """
-        args = Args(master_repo_names=MASTER_REPO_NAMES)
-        expected_ignore_tests = ["some", "tests"]
-        expected_hamcrest_path = "some/path/to/{}".format(_junit4_runner.HAMCREST_JAR)
-        expected_junit_path = "other/path/to/{}".format(_junit4_runner.JUNIT_JAR)
-        expected_rtd = RTD
-        expected_disable_security = False
-
-        junit4_hooks._ignore_tests = expected_ignore_tests
-        junit4_hooks._hamcrest_path = expected_hamcrest_path
-        junit4_hooks._junit_path = expected_junit_path
-        junit4_hooks._reference_tests_dir = expected_rtd
-        junit4_hooks._disable_security = expected_disable_security
-
-        junit4_hooks.parse_args(args)
-
-        assert junit4_hooks._ignore_tests == expected_ignore_tests
-        assert junit4_hooks._hamcrest_path == expected_hamcrest_path
-        assert junit4_hooks._junit_path == expected_junit_path
-        assert junit4_hooks._reference_tests_dir == expected_rtd
-        assert junit4_hooks._disable_security == expected_disable_security
-
-
-class TestConfigHook:
-    def test_with_full_config(
-        self, junit4_hooks, getenv_with_classpath, full_config_parser
-    ):
-        junit4_hooks.config_hook(full_config_parser)
-
-        assert junit4_hooks._hamcrest_path == HAMCREST_PATH
-        assert junit4_hooks._junit_path == JUNIT_PATH
-        assert junit4_hooks._classpath == CLASSPATH
-
-    def test_with_empty_config(self, junit4_hooks, getenv_with_classpath):
-        expected_junit = junit4_hooks._junit_path
-        expected_hamcrest = junit4_hooks._hamcrest_path
-
-        parser = ConfigParser()
-
-        junit4_hooks.config_hook(parser)
-
-        assert junit4_hooks._hamcrest_path == expected_hamcrest
-        assert junit4_hooks._junit_path == expected_junit
-        assert junit4_hooks._classpath == CLASSPATH
-
-
-class TestCloneParserHook:
-    @pytest.mark.parametrize("verbose", (None, False, True))
-    def test_arguments_get_added(self, junit4_hooks, verbose):
-        """Just test that `-rtd`, `-i`, `-ham` and `-junit` get added
-        correctly and that the args can then be parsed as expected.
-        """
-        parser = ArgumentParser()
-        sys_args = [
-            "-rtd",
-            RTD,
-            "-i",
-            " ".join(IGNORE_TESTS),
-            "-ham",
-            HAMCREST_PATH,
-            "-junit",
-            JUNIT_PATH,
-        ]
-
-        if verbose:
-            sys_args += ["-v"]
-
-        junit4_hooks.clone_parser_hook(parser)
-
-        args = parser.parse_args(sys_args)
-
-        assert args.reference_tests_dir == RTD
-        assert args.ignore_tests == IGNORE_TESTS
-        assert args.hamcrest_path == HAMCREST_PATH
-        assert args.junit_path == JUNIT_PATH
-        assert args.verbose == (False if verbose is None else verbose)
-
-    @pytest.mark.parametrize("skip_arg", ["ham", "junit", "rtd"])
-    def test_args_required_if_undefined(self, junit4_hooks, skip_arg):
-        """Test that junit, hamcrest and rtd args are required if they are not
-        defined in either config or classpath (for hamcrest and junit).
-        """
-        parser = ArgumentParser()
-        sys_args = ["-i", " ".join(IGNORE_TESTS)]
-
-        if skip_arg != "ham":
-            sys_args.extend(["-ham", HAMCREST_PATH])
-        if skip_arg != "junit":
-            sys_args.extend(["-junit", JUNIT_PATH])
-        if skip_arg != "rtd":
-            sys_args.extend(["-rtd", RTD])
-
-        junit4_hooks.clone_parser_hook(parser)
-
-        with pytest.raises(SystemExit):
-            parser.parse_args(sys_args)
-
-    def test_hamcrest_and_junit_not_required_if_on_classpath(
-        self, junit4_hooks, getenv_empty_classpath
-    ):
-        """Just checks that there is no chrash."""
-        config_parser = ConfigParser()
-        arg_parser = ArgumentParser()
-        sys_args = ["-rtd", RTD, "-i", " ".join(IGNORE_TESTS)]
-        getenv_empty_classpath.side_effect = (
-            lambda name: CLASSPATH_WITH_JARS if name == "CLASSPATH" else None
-        )
-
-        # this is where the classpath is picked up, still setup!
-        junit4_hooks.config_hook(config_parser)
-
-        # this is the actual test
-        junit4_hooks.clone_parser_hook(arg_parser)
-        arg_parser.parse_args(sys_args)  # should not crash!
-
-    def test_args_not_required_if_in_config(self, junit4_hooks, full_config_parser):
-        """Test that junit, hamcrest and rtd args are not required if they are
-        in the config.
-        """
-        args = Args(master_repo_names=MASTER_REPO_NAMES)
-        junit4_hooks.config_hook(full_config_parser)
-        parser = ArgumentParser()
-        junit4_hooks.clone_parser_hook(parser)
-
-        parser.parse_args([])  # should not crash
-
-
-def test_register():
-    """Just test that there is no crash"""
-    plug.manager.register(junit4)
